Refactor aiming into new aim controller
